move table account_activation into auth schema

+++ sql/functions/account/activate_account.sql
@@ -6,11 +6,11 @@ CREATE OR REPLACE FUNCTION activate_account( activation_string char(64)) RETURNS
     cur_activation RECORD;
   BEGIN
 
-    SELECT INTO cur_activation person_id, conference_id FROM account_activation WHERE account_activation.activation_string = activation_string AND account_creation > now() + '-1 day';
+    SELECT INTO cur_activation person_id, conference_id FROM auth.account_activation WHERE account_activation.activation_string = activation_string AND account_creation > now() + '-1 day';
     IF FOUND THEN
       INSERT INTO person_role(person_id, role_id) VALUES (cur_activation.person_id, (SELECT role_id FROM role WHERE tag = 'submitter'));
       INSERT INTO person_transaction( person_id, changed_when, changed_by, f_create ) VALUES (cur_activation.person_id, now(), cur_activation.person_id, 't');
-      DELETE FROM account_activation WHERE account_activation.activation_string = activation_string;
+      DELETE FROM auth.account_activation WHERE account_activation.activation_string = activation_string;
     ELSE
       RAISE EXCEPTION 'invalid activation string.';
     END IF;
